+++ app/src/androidTest/java/br/com/rbrthmn/ui/financialcompanion/FinancialCompanionScreenTest.kt
@@ -21,6 +21,7 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import br.com.rbrthmn.ui.financialcompanion.screens.FinancialCompanionScreen
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
